Update (미완) 평균 업글.py

+++ Simple things/(미완) 평균 업글.py
@@ -1,11 +1,12 @@
 #2022 / 5 / 30 ~ 미완
-#평균 구하는건데 총 몇갠지 모를 상황에서 쓸 수 있게 업글한거
+
+#평균 구하는건데 넣을 숫자가 총 몇갠지 모를 상황에서 쓸 수 있게 업글한거
 
 b = '0'
 while True:
 	if 'end' in b:
 		b.remove('end')
-		c = float(b)
+		c = int(b)
 		print(sum(c)/len(c))
 		break
 	else:
